requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.5
WTForms==3.0.1
email-validator==1.3.1
rapidfuzz
Unidecode==1.3.8
nltk
psycopg2-binary==2.9.9
gunicorn==20.1.0
Werkzeug==2.3.8
python-dotenv==1.0.0
Flask-Login==0.6.3
google-cloud-vision==3.7.2
beautifulsoup4

